Struct ic_crypto_tree_hash::HashTreeBuilderImpl

pub struct HashTreeBuilderImpl { /* fields omitted */ }

An implementation of the HashTreeBuilder-trait.

Implementations

impl HashTreeBuilderImpl

pub fn new() -> Self

pub fn as_hash_tree(&self) -> Option<HashTree>

Like into_hash_tree, but returns a copy of the hash tree. Does not panic!.

pub fn into_hash_tree(self) -> Option<HashTree>

Returns the HashTree corresponding to the traversed tree if the construction is complete, and None otherwise.

pub fn as_labeled_tree(&self) -> Option<LabeledTree<Digest>>

Returns the constructed LabeledTree if the construction is complete, and None otherwise. Does not panic!.