Trait ic_crypto_tree_hash::HashTreeBuilder

pub trait HashTreeBuilder {
    type WitnessGenerator;
    fn start_leaf(&mut self);

    fn write_leaf<T: AsRef<[u8]>>(&mut self, bytes: T);

    fn finish_leaf(&mut self);

    fn start_subtree(&mut self);

    fn new_edge<T: Into<Label>>(&mut self, label: T);

    fn finish_subtree(&mut self);

    fn witness_generator(&self) -> Option<Self::WitnessGenerator>;
}

HashTreeBuilder enables an iterative construction of a LabeledTree, which can also be accessed in form of a HashTree. The constructed LabeledTree is a part of the state of the Builder, and is build successively by adding leaves and subtrees. During the construction, the builder maintains an auxiliary state that describes the current position in the LabeledTree under construction. The auxiliary state is a list of nodes that corresponds to the path in the tree from the root to the current node being constructed. An example code to build the following labeled tree with a sub-tree and three leaves:

            root
            /  \
           /    \
     label_A    label_B
        /         \
       /           \
    "node A"       /\
                  /  \
                 /    \
            label_C   label_D
               /        \
              /          \
          "node C"     "node D is longer"

let builder = HashTreeBuilderImpl::new();

builder.start_subtree(); // start root

builder.new_edge("label_A");
builder.start_leaf();
builder.write_leaf("node A");
builder.finish_leaf();

builder.new_edge("label_B");

builder.start_subtree(); // start subtree (C, D)

builder.new_edge("label_C");
builder.start_leaf();
builder.write_leaf("node C");
builder.finish_leaf();

builder.new_edge("label_D");
builder.start_leaf();
builder.write_leaf("node D");
builder.write_leaf(" is longer");
builder.finish_leaf();

builder.finish_subtree(); // end subtree (C, D)
builder.finish_subtree(); // end root

Associated Types

type WitnessGenerator

Loading content...
Required methods

fn start_leaf(&mut self)

Starts a new leaf at the current position in the tree. Does not change the current position in the tree.

Panics

panics! if the current position is not an Undefined-node, i.e. if this builder is not in the initial state, or if the previous call was not new_edge().

fn write_leaf<T: AsRef<[u8]>>(&mut self, bytes: T)

Adds bytes to the content of the current leaf. Does not change the current position in the tree.

Panics

panics! if the current position is not a Leaf-node, i.e. if the previous call was neither start_leaf(), nor write_leaf().

fn finish_leaf(&mut self)

Finishes the current leaf. Changes the current position in the tree to the parent of the finished leaf (if any).

Panics

panics! if the current position is not a Leaf-node, i.e. if the previous call was neither start_leaf(), nor write_leaf().

fn start_subtree(&mut self)

Starts a new subtree at the current position in the tree. Does not change the current position in the tree.

Panics

panics! if the current position is not an Undefined-node, i.e. if this builder is not in the initial state, or if the previous call was not new_edge().

fn new_edge<T: Into<Label>>(&mut self, label: T)

Adds a new edge to the current subtree, and creates a new Undefined-node, to which the new edge leads. Changes the current position to the newly created node.

Panics

panics! if the current position is not a SubTree-node, i.e. if the previous call was neither start_subtree(), nor one of finish_leaf() or finish_subtree() that updated the current position to a parent that is a SubTree-node.

panics! if the current SubTree-node contains already an edge with the specified label.

fn finish_subtree(&mut self)

Finishes the current subtree. Changes the current position in the tree to the parent of the finished subtree (if any).

Panics

panics! if the current position is not a SubTree-node, i.e. if the previous call was neither start_subtree(), nor one of finish_leaf() or finish_subtree() that updated the current position to a parent that is a SubTree-node.

fn witness_generator(&self) -> Option<Self::WitnessGenerator>

Returns a WitnessGenerator for the constructed tree if the construction is complete, and None otherwise.

Does not panic!.

Loading content...