Trait ic_crypto_tree_hash::WitnessGenerator

pub trait WitnessGenerator {
    fn hash_tree(&self) -> &HashTree;

    fn witness(
        &self, 
        partial_tree: &LabeledTree<Vec<u8>>
    ) -> Result<Witness, TreeHashError>;

    fn mixed_hash_tree(
        &self, 
        partial_tree: &LabeledTree<Vec<u8>>
    ) -> Result<MixedHashTree, TreeHashError>;
}

WitnessGenerator is a container for a HashTree that offers access to the contained tree, and also provides operations related to the tree, like for example the computation of Witness-objects.

Required methods

fn hash_tree(&self) -> &HashTree

Returns the tree contained in this box.

Does not panic!.

fn witness(
&self,
partial_tree: &LabeledTree<Vec<u8>>
) -> Result<Witness, TreeHashError>

Computes and returns a witness for the paths specified partial_tree. All paths in partial_tree must end either with a Leaf-node or with an empty SubTree-node which in the original tree is also empty. If 'partial_tree' is inconsistent with the HashTree of this object, an error is returned.

Does not panic!.

fn mixed_hash_tree(
&self,
partial_tree: &LabeledTree<Vec<u8>>
) -> Result<MixedHashTree, TreeHashError>

Loading content...